refactor: move address_to_key to convert.py
